fix: a clamp lost to a shared call site, and four constants
RBEffectSizeSlider -sliderChangeWithTouchPoint: clamps a touch outside the
bar, and the reconstruction clamped it one way only. The compiler serves two
selectors from a single bl at 0x3be50: one arm loads barMin and jumps to it,
the other loads barMax and falls into it. Both arms had been collapsed into a
single barMax else, so a touch left of the bar jumped to the maximum. The
conversion is also fcvtas, round to nearest ties away, not the truncating cast
that was there, and the cmp/cinc/asr after it is an ordinary signed divide by
two rather than a hand-written shift.

The constant audit's four disagreements are three different failures.
kArtworkShadowOpacity was a wrong value at a right address: the slot holds 0.6
and the source said 0.15, two numbers one exponent step apart that share a
mantissa. kLineViewWhite and kTermsTextWhite were right values at wrong
addresses, both naming page 0x3be where the adrp says 0x2ee, and both wrong
addresses landed on real selector pointers rather than on nothing.
kProgressWidthInset was neither: the pool holds -60.0 and the code adds, while
the source declared +60.0 and subtracted. Identical arithmetic, but it left a
gate permanently red, which only teaches a reader to ignore it.

Also adds scan_property_types.py. scan_properties.py reads each attribute
string but compares only ownership and atomicity, so the leading T field, the
one record of a scalar's width and signedness, was parsed and discarded. It
found RBTutorialStatus backed by NSUInteger over a property and ivar the binary
encodes I: eight bytes declared where the shipped class kept four. Nothing else
in the tree shows that, since the enumeration's constants fit either width. The
scan covers 460 scalar properties, and was regression-tested by putting the
defect back, because a check that has never failed cannot be told apart from one
that cannot fail.

Worth recording separately: an ivar that read Q at eight bytes with a getter
plainly loading four turned out to be two correct readings of two different
binaries. Three extracted copies sit side by side and only two are the shipped
build. The address audit caught it by refusing to report 1786 of 1811 as
mismatched, which is not a tree full of defects but the wrong artefact.

Records the other nineteen methods of the batch that reported the slider, two
of them re-read here rather than taken on trust, and adds the missing
@ghidraAddress on RBHttpUtil's didWriteData: delegate method.

5188 of 6343 verified; 3507 from the mechanical passes and 1681 read by hand.

+++ Project/RBTutorialManager.h
@@ -27,7 +27,7 @@ NS_ASSUME_NONNULL_BEGIN
  * also written into the persisted @c RBUserSettingData tutorial-status map, plus the terminal
  * sentinel. @c RBTutorialStatusNone is the initial value assigned when the singleton is created.
  */
-typedef NS_ENUM(NSUInteger, RBTutorialStatus) {
+typedef NS_ENUM(unsigned int, RBTutorialStatus) {
     RBTutorialStatusMusicSelectStart = 0, /*!< First music-select walkthrough step; also the value
                                            *   below which @c +isTutorialMusicselect is true. */
     RBTutorialStatusPlayStart = 9,        /*!< The step at which the in-play walkthrough should
